Single Port Mode
Initial experimentation for multiplexing all client communication onto
the server's main listening port. Tracks connections with a map of
strings to channels. The string representation of the remote address is
used as the key. ACK, DATA, and ERRORs are sent on the connections
channel. Calls to conn.readFromNet read off the channel instead of a
network connection.